+++ tests/test_option_no_relative_x.sh
@@ -5,7 +5,7 @@ TEST_ROOT="$(dirname "$(realpath "$0")")"
 
 cool_mount() {
     # cd to use a relative path
-    (cd "$TEST_ROOT"; test_mount "$1" "$(fixture test.playlist)" -f "fixtures/fstab" -f "$(fixture script.sh)")
+    (cd "$TEST_ROOT"; test_mount "$@" "$(fixture test.playlist)" -f "fixtures/fstab" -f "$(fixture script.sh)")
 }
 
 run_test "--no-relative mount" cool_mount --no-relative 
@@ -14,7 +14,7 @@ subtest "Relative path in a list is ignored" test ! -f "$TEST_MOUNT_POINT/test.p
 subtest "Absolute path in --file is respected" test -f "$TEST_MOUNT_POINT/script.sh"
 subtest "Relative path in --file is ignored" compare_file_info "$TEST_MOUNT_POINT/fstab" "/etc/fstab"
 
-run_test "-n mount" cool_mount -n
+run_test "-N mount" cool_mount -N
 subtest "Absolute path in a list is respected" test -f "$TEST_MOUNT_POINT/hosts"
 subtest "Relative path in a list is ignored" test ! -f "$TEST_MOUNT_POINT/test.playlist"
 subtest "Absolute path in --file is respected" test -f "$TEST_MOUNT_POINT/script.sh"
@@ -26,8 +26,24 @@ subtest "Relative path in a list is respected" test -f "$TEST_MOUNT_POINT/test.p
 subtest "Absolute path in --file is respected" test -f "$TEST_MOUNT_POINT/script.sh"
 subtest "Relative path in --file is ignored" compare_file_info "$TEST_MOUNT_POINT/fstab" "/etc/fstab"
 
-run_test "--no-relative-lists mount" cool_mount --no-relative-lists
+run_test "--no-relative-paths mount" cool_mount --no-relative-paths
 subtest "Absolute path in a list is respected" test -f "$TEST_MOUNT_POINT/hosts"
 subtest "Relative path in a list is ignored" test ! -f "$TEST_MOUNT_POINT/test.playlist"
 subtest "Absolute path in --file is respected" test -f "$TEST_MOUNT_POINT/script.sh"
 subtest "Relative path in --file is respected" compare_file_info "$TEST_MOUNT_POINT/fstab" "$(fixture fstab)"
+
+run_test "--no-relative --relative-files mount" cool_mount --no-relative --relative-files
+subtest "Relative path in a list is ignored" test ! -f "$TEST_MOUNT_POINT/test.playlist"
+subtest "Relative path in --file is respected" compare_file_info "$TEST_MOUNT_POINT/fstab" "$(fixture fstab)"
+
+run_test "--no-relative --relative-paths mount" cool_mount --no-relative --relative-paths
+subtest "Relative path in a list is respected" test -f "$TEST_MOUNT_POINT/test.playlist"
+subtest "Relative path in --file is ignored" compare_file_info "$TEST_MOUNT_POINT/fstab" "/etc/fstab"
+
+run_test "--no-relative --relative mount" cool_mount --no-relative --relative
+subtest "Relative path in a list is respected" test -f "$TEST_MOUNT_POINT/test.playlist"
+subtest "Relative path in --file is respected" compare_file_info "$TEST_MOUNT_POINT/fstab" "$(fixture fstab)"
+
+run_test "--relative --no-relative mount" cool_mount --relative --no-relative
+subtest "Relative path in a list is ignored" test ! -f "$TEST_MOUNT_POINT/test.playlist"
+subtest "Relative path in --file is ignored" compare_file_info "$TEST_MOUNT_POINT/fstab" "/etc/fstab"