+++ tests/test_option_symlink.sh
@@ -3,7 +3,7 @@
 TEST_ROOT="$(dirname "$(realpath "$0")")"
 . "$TEST_ROOT/setup.sh"
 
-run_test "--symlink mount" test_mount --symlink "$(fixture test.playlist)"
+run_test "--symlinks mount" test_mount --symlinks "$(fixture test.playlist)"
 subtest "File is a symlink" test "q$(extract_mode "$TEST_MOUNT_POINT/test.playlist")" = "qlrwxrwxrwx"
 pending subtest "Symlink is created today" test "q$(extract_file_info "$TEST_MOUNT_POINT/test.playlist" | cut -d' ' -f6)" = "q$(date -I)"